Move properties into separate package
This cleanup makes working with them more convenient.
