+++ src/main/java/de/thomas_oster/liblasercut/VectorPart.java
@@ -18,6 +18,8 @@
  */
 package de.thomas_oster.liblasercut;
 
+import de.thomas_oster.liblasercut.properties.LaserProperty;
+
 import java.util.LinkedList;
 import java.util.List;
