+++ src/main/java/de/thomas_oster/liblasercut/drivers/Dummy.java
@@ -22,8 +22,8 @@ import de.thomas_oster.liblasercut.IllegalJobException;
 import de.thomas_oster.liblasercut.JobPart;
 import de.thomas_oster.liblasercut.LaserCutter;
 import de.thomas_oster.liblasercut.LaserJob;
-import de.thomas_oster.liblasercut.PowerSpeedFocusFrequencyProperty;
-import de.thomas_oster.liblasercut.PowerSpeedFocusProperty;
+import de.thomas_oster.liblasercut.properties.PowerSpeedFocusFrequencyProperty;
+import de.thomas_oster.liblasercut.properties.PowerSpeedFocusProperty;
 import de.thomas_oster.liblasercut.ProgressListener;
 import de.thomas_oster.liblasercut.ProgressListenerDummy;
 import de.thomas_oster.liblasercut.Raster3dPart;
