+++ src/main/java/de/thomas_oster/liblasercut/drivers/EpilogCutter.java
@@ -26,7 +26,10 @@ package de.thomas_oster.liblasercut.drivers;
 
 import de.thomas_oster.liblasercut.*;
 import de.thomas_oster.liblasercut.platform.Point;
-import de.thomas_oster.liblasercut.platform.Util;
+import de.thomas_oster.liblasercut.properties.LaserProperty;
+import de.thomas_oster.liblasercut.properties.PowerSpeedFocusFrequencyProperty;
+import de.thomas_oster.liblasercut.properties.PowerSpeedFocusProperty;
+
 import java.io.*;
 import java.net.InetSocketAddress;
 import java.net.Socket;
