+++ src/main/java/de/thomas_oster/liblasercut/drivers/GenericGcodeDriver.java
@@ -18,12 +18,12 @@
  */
 package de.thomas_oster.liblasercut.drivers;
 
-import de.thomas_oster.liblasercut.FloatPowerSpeedFocusProperty;
+import de.thomas_oster.liblasercut.properties.FloatPowerSpeedFocusProperty;
 import de.thomas_oster.liblasercut.IllegalJobException;
 import de.thomas_oster.liblasercut.JobPart;
 import de.thomas_oster.liblasercut.LaserCutter;
 import de.thomas_oster.liblasercut.LaserJob;
-import de.thomas_oster.liblasercut.LaserProperty;
+import de.thomas_oster.liblasercut.properties.LaserProperty;
 import de.thomas_oster.liblasercut.OptionSelector;
 import de.thomas_oster.liblasercut.ProgressListener;
 import de.thomas_oster.liblasercut.ProgressListenerDummy;
