+++ src/main/java/de/thomas_oster/liblasercut/drivers/SampleDriver.java
@@ -23,8 +23,8 @@ import de.thomas_oster.liblasercut.IllegalJobException;
 import de.thomas_oster.liblasercut.JobPart;
 import de.thomas_oster.liblasercut.LaserCutter;
 import de.thomas_oster.liblasercut.LaserJob;
-import de.thomas_oster.liblasercut.LaserProperty;
-import de.thomas_oster.liblasercut.PowerSpeedFocusFrequencyProperty;
+import de.thomas_oster.liblasercut.properties.LaserProperty;
+import de.thomas_oster.liblasercut.properties.PowerSpeedFocusFrequencyProperty;
 import de.thomas_oster.liblasercut.ProgressListener;
 import de.thomas_oster.liblasercut.VectorCommand;
 import de.thomas_oster.liblasercut.VectorPart;
