+++ src/main/java/de/thomas_oster/liblasercut/examples/PhotoPrint.java
@@ -22,8 +22,8 @@ import de.thomas_oster.liblasercut.BlackWhiteRaster;
 import de.thomas_oster.liblasercut.BlackWhiteRaster.DitherAlgorithm;
 import de.thomas_oster.liblasercut.IllegalJobException;
 import de.thomas_oster.liblasercut.LaserJob;
-import de.thomas_oster.liblasercut.PowerSpeedFocusFrequencyProperty;
-import de.thomas_oster.liblasercut.PowerSpeedFocusProperty;
+import de.thomas_oster.liblasercut.properties.PowerSpeedFocusFrequencyProperty;
+import de.thomas_oster.liblasercut.properties.PowerSpeedFocusProperty;
 import de.thomas_oster.liblasercut.RasterPart;
 import de.thomas_oster.liblasercut.VectorPart;
 import de.thomas_oster.liblasercut.drivers.EpilogZing;
