+++ src/main/java/de/thomas_oster/liblasercut/properties/FloatMinMaxPowerSpeedFrequencyProperty.java
@@ -18,7 +18,7 @@
  *
  **/
 
-package de.thomas_oster.liblasercut;
+package de.thomas_oster.liblasercut.properties;
 
 import java.util.Arrays;
 import java.util.LinkedList;
