+++ src/main/java/de/thomas_oster/liblasercut/properties/FloatPowerSpeedFocusProperty.java
@@ -16,7 +16,7 @@
   along with LibLaserCut. If not, see <http://www.gnu.org/licenses/>.
 
  */
-package de.thomas_oster.liblasercut;
+package de.thomas_oster.liblasercut.properties;
 
 /**
  * The LaserProperty holds all the parameters for parts of the LaserJob.
@@ -24,15 +24,14 @@ package de.thomas_oster.liblasercut;
  *
  * @author oster
  */
-public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
+public class FloatPowerSpeedFocusProperty implements LaserProperty
 {
 
   private float power = 0;
   private float speed = 100;
   private float focus = 0;
-  private int frequency = 500;
 
-  public FloatPowerSpeedFocusFrequencyProperty()
+  public FloatPowerSpeedFocusProperty()
   {
   }
 
@@ -40,7 +39,6 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
    * Sets the Laserpower. Valid values are from 0 to 100.
    * In 3d-Raster mode, the intensity is scaled to this power setting
    */
-  @Override
   public void setPower(float power)
   {
     power = power < 0 ? 0 : power;
@@ -48,7 +46,6 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
     this.power = power;
   }
 
-  @Override
   public float getPower()
   {
     return power;
@@ -64,7 +61,6 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
     this.speed = speed;
   }
 
-  @Override
   public float getSpeed()
   {
     return speed;
@@ -91,29 +87,18 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
   {
     return this.focus;
   }
-  
-  public void setFrequency(int f)
-  {
-    this.frequency = f;
-  }
-  
-  public int getFrequency()
-  {
-    return this.frequency;
-  }
 
   @Override
-  public FloatPowerSpeedFocusFrequencyProperty clone()
+  public FloatPowerSpeedFocusProperty clone()
   {
-    FloatPowerSpeedFocusFrequencyProperty p = new FloatPowerSpeedFocusFrequencyProperty();
+    FloatPowerSpeedFocusProperty p = new FloatPowerSpeedFocusProperty();
     p.focus = focus;
-    p.frequency = frequency;
     p.power = power;
     p.speed = speed;
     return p;
   }
 
-  private static final String[] propertyNames = new String[]{"power", "speed", "focus", "frequency"};
+  private static final String[] propertyNames = new String[]{"power", "speed", "focus"};
   
   @Override
   public String[] getPropertyKeys()
@@ -136,10 +121,6 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
     {
       return this.getFocus();
     }
-    else if ("frequency".equals(name))
-    {
-      return this.getFrequency();
-    }
     return null;
   }
 
@@ -157,10 +138,6 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
     else if ("focus".equals(name))
     {
       this.setFocus((Float) value);
-    }
-    else if ("frequency".equals(name))
-    {
-      this.setFrequency((Integer) value);
     }
     else
     {
@@ -232,17 +209,14 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
         if (getClass() != obj.getClass()) {
             return false;
         }
-        final FloatPowerSpeedFocusFrequencyProperty other = (FloatPowerSpeedFocusFrequencyProperty) obj;
+        final FloatPowerSpeedFocusProperty other = (FloatPowerSpeedFocusProperty) obj;
         if (Float.floatToIntBits(this.power) != Float.floatToIntBits(other.power)) {
             return false;
         }
         if (Float.floatToIntBits(this.speed) != Float.floatToIntBits(other.speed)) {
             return false;
         }
-        if (Float.floatToIntBits(this.focus) != Float.floatToIntBits(other.focus)) {
-            return false;
-        }
-        return this.frequency == other.frequency;
+        return Float.floatToIntBits(this.focus) == Float.floatToIntBits(other.focus);
     }
 
     @Override
@@ -251,7 +225,6 @@ public class FloatPowerSpeedFocusFrequencyProperty implements LaserProperty
         hash = 67 * hash + Float.floatToIntBits(this.power);
         hash = 67 * hash + Float.floatToIntBits(this.speed);
         hash = 67 * hash + Float.floatToIntBits(this.focus);
-        hash = 67 * hash + this.frequency;
         return hash;
     }
