+++ src/main/java/de/thomas_oster/liblasercut/properties/LaserProperty.java
@@ -16,7 +16,9 @@
   along with LibLaserCut. If not, see <http://www.gnu.org/licenses/>.
 
  */
-package de.thomas_oster.liblasercut;
+package de.thomas_oster.liblasercut.properties;
+
+import de.thomas_oster.liblasercut.Customizable;
 
 /**
  * The LaserProperty holds all the parameters for parts of the LaserJob.
