+++ src/test/java/de/thomas_oster/liblasercut/drivers/EpilogHelixCutterTest.java
@@ -19,8 +19,8 @@
 package de.thomas_oster.liblasercut.drivers;
 
 import de.thomas_oster.liblasercut.*;
-import de.thomas_oster.liblasercut.platform.Point;
 import de.thomas_oster.liblasercut.platform.Util;
+import de.thomas_oster.liblasercut.properties.LaserProperty;
 import org.junit.Test;
 
 import java.util.Arrays;
